% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessAccuracy.R
\name{assess.accuracy}
\alias{assess.accuracy}
\title{Function to assess the accuracy criterion for CPUE and abundance}
\usage{
assess.accuracy(ts_mat, N_vec, sig_vec, q_bar, sd_q, acc_lev, A, n_mc = 100)
}
\arguments{
\item{ts_mat}{A matrix specifying the transect/set arrangement. Columns represent transects while rows represent sets. Eg: cbind(c(T,T,T,F),c(T,T,F,F),c(T,T,T,T))}

\item{N_vec}{A numerical vector specifying the "true" mean abundance}

\item{sig_vec}{A numerical vector specifying the between transect log scale standard deviation}

\item{q_bar}{The estimate of the catchability coefficient}

\item{sd_q}{An estimate of the standard deviation of the catchability coefficient}

\item{acc_lev}{A vector of length two giving the prescribed levels of accuracy for CPUE and N respectively.}

\item{A}{The area of the lake (in XX)}

\item{n_mc}{The number of times to run the Monte Carlo simulation}
}
\value{
Returns a matrix giving the percentage of the time that the estimate is within acc_lev percent of the true value. The rows of the matrix coorespond to the entries in N_vec while the columns coorespond to entries in sig_vec.
}
\description{
Function to assess the accuracy criterion for CPUE and abundance
}
\examples{
load("S:/Jordy/louiseOP2020/Data/lakelines.Rdata") # loading ts_mat
N <- c(1000, 5000, 10000) 
sig_vec <- seq(0.1, 1, by=.1)
q_bar <- 0.6478999
sd_q <- 0.07966257
acc_lev <- c(0.3, 0.3)
A <- 6519
assess.accuracy(ts_mat, N_vec, sig_vec, q_bar, sd_q, acc_lev, A)
}
