% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCtMat.R
\name{get.ct.mat}
\alias{get.ct.mat}
\title{Function to read in and prepare the 200 count data}
\usage{
get.ct.mat(file, sheet)
}
\arguments{
\item{file}{The xlx file to read in. The file should consist of three columns where the first column lists contains the transect #, the second column contains the set within transect #, and the third column contains the fish count. See as.data.frame(read_excel("S:/Jordy/louiseOP2020/Data/CPUE.xls", col_names = F, sheet = "CPUEL"))}

\item{sheet}{The name of the sheet to read in.}
}
\value{
Returns a matrix containing the 2005 count data. Note: columns represent transects and rows represent sets within transects.
}
\description{
Function to read in and prepare the 200 count data
}
\examples{
 get.ct.mat("S:/Jordy/louiseOP2020/Data/CPUE.xls", "CPUEL")
}
