% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthCompSS.R
\name{length.comp.ss}
\alias{length.comp.ss}
\title{Function to calculate sample size requirements for length composition estimates using the methods of Thompson 1987}
\usage{
\method{length}{comp.ss}(alpha, dist)
}
\value{
returns the minimum sample size to achieve the desired level of precision at the specified significance level
}
\description{
Function to calculate sample size requirements for length composition estimates using the methods of Thompson 1987
}
\examples{
length.comp.ss(0.05, 0.10) # To estimate the proportions within 10 percentage points of the actual values 95\% of the time
}
