% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickCoords.R
\name{pick.coords}
\alias{pick.coords}
\title{Function to get the coordinate locations of sets for the 2020 study}
\usage{
pick.coords(lakelines, n_sets, seed = 123)
}
\arguments{
\item{lakelines}{SpatialPolygons object representing the sampling area}

\item{n_sets}{The total number of sets}

\item{seed}{Sets a seed for reproducability}
}
\value{
Prints a map of the selected set locations and returns a list containing information related to the selected transects. $ts_mat will be used in assessAccuracy() and calcPower(); $table and $setcoords will be given to project biologists to position the sets.
}
\description{
Function to get the coordinate locations of sets for the 2020 study
}
\examples{
load("S:/Jordy/louiseOP2020/Data/lakelines.Rdata") # Load Lake Louise lakelines
nmax <- 491
seed <- set.seed(123)
pick.coords(lakelines, n_max, seed)
}
